+++ js/src/components/courses/Card.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 import { colors } from "yoast-components";
-import Banner from "./Banner";
+import Banner from "./CardBanner";
 
 const Container = styled.div`
 	position: relative;
