+++ js/src/courses-overview.js
@@ -1,7 +1,7 @@
 import { render } from "@wordpress/element";
 import styled from "styled-components";
 import { FullHeightCard } from "./components/courses/Card";
-import CourseDetails from "./components/courses/CourseDetails";
+import CardDetails from "./components/courses/CardDetails";
 
 const element = document.getElementById( "yoast-courses-overview" );
 
@@ -37,60 +37,9 @@ const courses = [
 		description: "Do you know the essential first step of good SEO? It’s keyword research. In this training, you’ll learn how to research and select the keywords that will guide searchers to your pages.",
 		courseUrl: "https://yoast.com/academy/keyword-research-training/",
 		shopUrl: "https://yoast.com/cart/?add-to-cart=1311259",
-	},
-	{
-		title: "Structured Data",
-		image: "../wp-content/plugins/wordpress-seo/js/src/components/courses/sample_course_card_header.png",
-		description: "Get your site ready for rich search results that bring in loads of customers. Make your site stand out from the crowd by adding structured data!",
-		courseUrl: "https://yoast.com/academy/keyword-research-training/",
-		shopUrl: "https://yoast.com/cart/?add-to-cart=1311259",
 		banner: {
 			text: "Free trial available",
 		},
-	},
-	{
-		title: "SEO for Beginners",
-		image: "../wp-content/plugins/wordpress-seo/js/src/components/courses/sample_course_card_header.png",
-		description: "In this free course, you'll get quick wins to make your site rank higher in Google, Bing, and Yahoo.",
-		courseUrl: "https://yoast.com/academy/keyword-research-training/",
-		shopUrl: "https://yoast.com/cart/?add-to-cart=1311259",
-		banner: {
-			text: "Free",
-		},
-	},
-	{
-		title: "Keyword Research",
-		image: "../wp-content/plugins/wordpress-seo/js/src/components/courses/sample_course_card_header.png",
-		description: "Do you know the essential first step of good SEO? It’s keyword research. In this training, you’ll learn how to research and select the keywords that will guide searchers to your pages.",
-		courseUrl: "https://yoast.com/academy/keyword-research-training/",
-		shopUrl: "https://yoast.com/cart/?add-to-cart=1311259",
-	},
-	{
-		title: "Structured Data",
-		image: "../wp-content/plugins/wordpress-seo/js/src/components/courses/sample_course_card_header.png",
-		description: "Get your site ready for rich search results that bring in loads of customers. Make your site stand out from the crowd by adding structured data!",
-		courseUrl: "https://yoast.com/academy/keyword-research-training/",
-		shopUrl: "https://yoast.com/cart/?add-to-cart=1311259",
-		banner: {
-			text: "Free trial available",
-		},
-	},
-	{
-		title: "SEO for Beginners",
-		image: "../wp-content/plugins/wordpress-seo/js/src/components/courses/sample_course_card_header.png",
-		description: "In this free course, you'll get quick wins to make your site rank higher in Google, Bing, and Yahoo.",
-		courseUrl: "https://yoast.com/academy/keyword-research-training/",
-		shopUrl: "https://yoast.com/cart/?add-to-cart=1311259",
-		banner: {
-			text: "Free",
-		},
-	},
-	{
-		title: "Keyword Research",
-		image: "../wp-content/plugins/wordpress-seo/js/src/components/courses/sample_course_card_header.png",
-		description: "Do you know the essential first step of good SEO? It’s keyword research. In this training, you’ll learn how to research and select the keywords that will guide searchers to your pages.",
-		courseUrl: "https://yoast.com/academy/keyword-research-training/",
-		shopUrl: "https://yoast.com/cart/?add-to-cart=1311259",
 	},
 	{
 		title: "Structured Data",
@@ -122,7 +71,7 @@ render(
 					header={ getHeaderData( course ) }
 					banner={ course.banner }
 				>
-					<CourseDetails
+					<CardDetails
 						title={ course.title }
 						description={ course.description }
 						courseUrl={ course.courseUrl }
